+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/service/JobInstanceTaskService.java
@@ -3,9 +3,11 @@ package io.openjob.server.admin.service;
 import io.openjob.server.admin.request.task.ListChildTaskRequest;
 import io.openjob.server.admin.request.task.ListTaskRequest;
 import io.openjob.server.admin.request.task.ListTaskLogRequest;
+import io.openjob.server.admin.request.task.StopTaskRequest;
 import io.openjob.server.admin.vo.task.ListChildTaskVO;
 import io.openjob.server.admin.vo.task.ListTaskVO;
 import io.openjob.server.admin.vo.task.ListTaskLogVO;
+import io.openjob.server.admin.vo.task.StopTaskVO;
 import io.openjob.server.common.vo.PageVO;
 
 /**
@@ -30,6 +32,14 @@ public interface JobInstanceTaskService {
      * @return PageVO
      */
     PageVO<ListChildTaskVO> getChildList(ListChildTaskRequest request);
+
+    /**
+     * Stop task
+     *
+     * @param request request
+     * @return StopTaskVO
+     */
+    StopTaskVO stopTask(StopTaskRequest request);
 
     /**
      * Get task log list
