+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/vo/task/ListTaskVO.java
@@ -21,6 +21,9 @@ public class ListTaskVO {
     @ApiModelProperty(value = "Circle id")
     private Long circleId;
 
+    @ApiModelProperty(value = "Dispatch version")
+    private Long dispatchVersion;
+
     @ApiModelProperty(value = "Task id")
     private String taskId;
