+++ openjob-server/openjob-server-admin/src/main/java/io/openjob/server/admin/vo/task/StopTaskVO.java
@@ -1,0 +1,11 @@
+package io.openjob.server.admin.vo.task;
+
+import lombok.Data;
+
+/**
+ * @author stelin swoft@qq.com
+ * @since 1.0.7
+ */
+@Data
+public class StopTaskVO {
+}
