+++ openjob-server/openjob-server-scheduler/src/main/java/io/openjob/server/scheduler/dto/StopTaskResponseDTO.java
@@ -1,0 +1,11 @@
+package io.openjob.server.scheduler.dto;
+
+import lombok.Data;
+
+/**
+ * @author stelin swoft@qq.com
+ * @since 1.0.7
+ */
+@Data
+public class StopTaskResponseDTO {
+}
